+++ config/common/src/main/java/band/kessoku/lib/api/config/api/Codec.java
@@ -1,4 +1,19 @@
-package band.kessoku.lib.config.api;
+/*
+ * Copyright (c) 2024 KessokuTeaTime
+ *
+ * Licensed under the GNU Lesser General Pubic License, Version 3 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.gnu.org/licenses/lgpl-3.0.html
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package band.kessoku.lib.api.config.api;
 
 /**
  * {@code Codec} is the key part of data encoding. <br>
@@ -9,6 +24,8 @@ package band.kessoku.lib.config.api;
  *
  * @see ConfigValue
  * @param <T> The type we will encode and decode.
+ *
+ * @author AmarokIce
  */
 public interface Codec<T> {
     /**
