+++ config/common/src/main/java/band/kessoku/lib/api/config/api/Comment.java
@@ -1,5 +1,20 @@
-package band.kessoku.lib.config.api;
+package band.kessoku.lib.api.config.api;
 
+/*
+ * Copyright (c) 2024 KessokuTeaTime
+ *
+ * Licensed under the GNU Lesser General Pubic License, Version 3 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.gnu.org/licenses/lgpl-3.0.html
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -28,6 +43,8 @@ import java.lang.annotation.Target;
  * {@code someoneField = test,}
  *
  * @see Config @Config
+ *
+ * @author AmarokIce
  */
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
