+++ config/common/src/main/java/band/kessoku/lib/api/config/api/ConfigValue.java
@@ -1,7 +1,22 @@
-package band.kessoku.lib.config.api;
+/*
+ * Copyright (c) 2024 KessokuTeaTime
+ *
+ * Licensed under the GNU Lesser General Pubic License, Version 3 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.gnu.org/licenses/lgpl-3.0.html
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package band.kessoku.lib.api.config.api;
 
-import band.kessoku.lib.config.ConfigHandler;
-import band.kessoku.lib.config.values.*;
+import band.kessoku.lib.api.config.ConfigHandler;
+import band.kessoku.lib.api.config.values.*;
 
 import java.util.Objects;
 
@@ -11,7 +26,9 @@ import java.util.Objects;
  *
  * @see Codec
  * @see ConfigValue#ConfigValue(Codec, Object) 
- * @param <T>
+ * @param <T> The value should record.
+ *
+ * @author AmarokIce
  */
 public abstract class ConfigValue<T> {
     public final Codec<T> codec;
