+++ config/common/src/main/java/band/kessoku/lib/api/config/api/package-info.java
@@ -1,19 +1,12 @@
-package band.kessoku.lib.config.api;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
 /**
- * By default, the {@code ConfigHandler} takes the field name as the name of the config field, but you can use this annotation to commit a new name.
+ * The API for config. It all started in {@link band.kessoku.lib.api.config.api.Config @Config}.<br>
  * <br>
  * {@snippet :
  * import band.kessoku.lib.config.api.config.Name;
  * import band.kessoku.lib.config.api.config.Comment;
  * import band.kessoku.lib.config.values.config.StringValue;
  *
- * @Config("mymodid")
+ * @Config("mymodid)
  * public class MyConfig {
  *      @Comment({"First comment", "Second comment"})
  *      @Name("someoneField")
@@ -26,10 +19,5 @@ import java.lang.annotation.Target;
  * {@code // Second comment} <br>
  * {@code someoneField = test,}
  *
- * @see Config @Config
  */
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Name {
-    String value();
-}
+package band.kessoku.lib.api.config.api;
