+++ config/common/src/main/java/band/kessoku/lib/api/config/values/IdentifierValue.java
@@ -1,12 +1,30 @@
-package band.kessoku.lib.config.values;
+/*
+ * Copyright (c) 2024 KessokuTeaTime
+ *
+ * Licensed under the GNU Lesser General Pubic License, Version 3 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.gnu.org/licenses/lgpl-3.0.html
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package band.kessoku.lib.api.config.values;
 
-import band.kessoku.lib.config.api.Codec;
-import band.kessoku.lib.config.api.ConfigValue;
+import band.kessoku.lib.api.config.api.Codec;
+import band.kessoku.lib.api.config.api.ConfigValue;
 import net.minecraft.block.Block;
 import net.minecraft.item.Item;
 import net.minecraft.registry.Registries;
 import net.minecraft.util.Identifier;
 
+/**
+ * @author AmarokIce
+ */
 public class IdentifierValue extends ConfigValue<Identifier> {
     public static final Codec<Identifier> CODEC = new Codec<>() {
         @Override
