--- config/common/src/main/java/band/kessoku/lib/config/KessokuConfig.java
@@ -1,10 +1,0 @@
-package band.kessoku.lib.config;
-
-import org.slf4j.Marker;
-import org.slf4j.MarkerFactory;
-
-public class KessokuConfig {
-    public static final String MOD_ID = "kessoku_config";
-    public static final String NAME = "Kessoku Config API";
-    public static final Marker MARKER = MarkerFactory.getMarker("[" + NAME + "]");
-}
