--- config/common/src/main/java/band/kessoku/lib/config/package-info.java
@@ -1,6 +1,0 @@
-/**
- * You can use annotations to create a config and save it to any file format you like. <br>
- * @see band.kessoku.lib.config.api.Config Config
- * @see band.kessoku.lib.config.api.ConfigValue ConfigValue
- */
-package band.kessoku.lib.config;
