--- config/common/src/main/java/band/kessoku/lib/config/values/package-info.java
@@ -1,7 +1,0 @@
-/**
- * The default values for config. <br>
- * For any customize value, extends the {@link band.kessoku.lib.config.api.ConfigValue ConfigValue}
- * and create your {@link band.kessoku.lib.config.api.Codec Codec}. <br>
- * After that, you can use it without additional registration or anything else.
- */
-package band.kessoku.lib.config.values;
