+++ settings.gradle
@@ -20,7 +20,7 @@ includeModule("registry") // Registry
 includeModule("lifecycle-events") // Lifecycle Events
 includeModule("command") // Command API
 includeModule("keybinding") // Keybinding API
-// includeModule("config") // Config API
+includeModule("config") // Config API
 includeModule("entrypoint") // Entrypoint API
 includeModule("entity-events") // Entity Events
 includeModule("blockentity") // Block Entity
